In Python's world "float"s are 64 bits wide

+++ test/test_fetch_df.py
@@ -46,7 +46,7 @@ class TestClass:
         assert df["count"].dtype == np.int32
 
         df = f.fetch("chr2R:10,000,000-15,000,000", count_type="float").to_df()
-        assert df["count"].dtype == np.float32
+        assert df["count"].dtype == np.float64
 
         df = f.fetch("chr2R\t10000000\t15000000", query_type="BED").to_df()
         assert len(df) == 1275
@@ -77,7 +77,7 @@ class TestClass:
         assert df["count"].dtype == np.int32
 
         df = f.fetch("chr2R:10,000,000-15,000,000", "chrX:0-10,000,000", count_type="float").to_df()
-        assert df["count"].dtype == np.float32
+        assert df["count"].dtype == np.float64
 
         df = f.fetch("chr2R\t10000000\t15000000", "chrX\t0\t10000000", query_type="BED").to_df()
         assert len(df) == 4995
